Initial commit
This commit comes quite late, as most of the project had been handled by
the VC of the accompanying paper. As a result much of the implementation
is already present (as far as the paper is concerned), with a final hole
for the completeness proof. This hole has been shown to be unfillable,
as discussed in the paper, by proving bottom with the proof of
completeness.