Rearrange modules to emphasise co/inductive split
The definition of the coinductive interpretation of types was mixed
together with both the definition of Value and the semantics of the
types. In order to emphasise the distinction between the inductive and
coinductive interpretations, the definitions related to the inductive
form are in a containing module Inductive, and the definitions related
to the coinductive form are in a containing module Coinductive. Value is
defined in its own module, as it has no dependency on the types.